src/requirements.txt
hatterbot
chatterbot_corpus
slackclient
websocket-client